+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.0] - 2025-04/14
+
+### Added
+
+- Type hints for the `HSIClasso` class
+
+### Changed
+
+- `pyhsiclasso.nlars.nlars` now returns an numpy array for `A` instead of a list of ints`
+- Removed file input methods. No need to have import functions here, just import them yourself.
+
 ## [1.13.0] - 2024-08-23
 
 ### Added